Began Project, Created Three Problems
Dean Woelfle
Dec 16, 2020
Start time: 9:45 
End time: 11:01 

Began project by choosing to do the Euler problems. I saw how I will have to create code for answering many different problems, so I coded a simple mechanism to let the user choose which problem they want to see. I may improve on this mechanic in the future, but for now, it works well. The first problem was to find the sum of all the multiples of 3 or 5 below 1000. I began this by creating a for loop that runs through all the numbers between 1 and 1000. For each number, I divided it by 3 and then 5 and saw if the remainder for either was equal to 0. If either remainder was equal to 0, I added the number to the total sum. Once I finally hit 1000 and have run through every number, I print out the sum as it is the final answer. For the second problem, I needed to find the sum of all the even valued terms less than four million in the Fibonacci sequence. I began by creating a list for our Fibonacci sequence numbers, starting it off with 1 and 2 as its values. I then created a while loop that only runs the next line of code if the last and second to last entry in our Fibonacci list add up to less than four million, to make sure that our values don't go over four million, as per the rules. If these two numbers aren't over four million, I add them together to create the next number to be added to our sequence. Once this is done, our list of Fibonacci numbers is complete. I then run through every number in this list and divide it by two. If this has a remainder of 0, it is even and is therefore added to the sum. This sum is then converted into a string and printed out, as it is our final answer. For the third problem, I needed to find the largest prime factor of the number 600851475143. I first created a list for the prime factors I will soon calculate. Then, I ran through a list of potential factors of 600851475143 and saw whether the remainder of the two is equal to 0, as this would signify that the number is a factor. For each factor, I divided it by all their potential factors. If any of the remainders, besides the one after being divided by 1 or itself, are equal to 0, then the number is not prime. If it is prime, I add it to the list of prime factors. At the end, I print off the most recent, largest prime factor as my answer.