Added problems #4 - #6
Once again, I have completed three more problems for this Lab. Additionally, I tweaked some of the pre-existing code to make it better and more efficient. My code is filled with comments, so it would be kind of repetitive for me to redescribe what it does. All my code so far works very well.