Added #9 and #10
Just as last time, I spent my 75 min working on two problems, being 9 and 10. 9 was confusing at first and I had to spend some time figuring out how to do it. Overall, not too challenging, and my method and work is explained in my code. Once I started 10, I realized how similar it was to #7 as they both required finding primes. As a result, I recycled much of the code in 7 for 10 and it worked well. My process and work is in my code.