Completed problem 11
Initially, I had a lot of issues with 11, resulting in me only completing this single problem. I was really confused on how to use the grid and how I could make my code read a grid. This was a rabbit hole that I realized I would never be able to solve, as I didn't know enough about Python to compute grids. Then I realized that if I think about the grid less like a grid and more like a single list of 400 consecutive characters ARRANGED in a 20 x 20 grid, I would be able to calculate my answers much easier. While a horizontal line is moving 1 across, a verticle line is moving 20 across each time as its going on to the next line, with diagonal lines skipping either 19 or 21 spaces. This is explained further in my docstring. Worked immediately and much faster/better than I honestly expected.