Added 12 and 13
Both 12 and 13 weren't very complicated. Once a read the problems and thought about them for a minute I knew what I needed to do for each problem. 12 is just while loops and if loops adding up numbers, checking all their possible divisors, and checking if there are over 500. It takes a significant amount of time though, as it is running through so many numbers, but it works. 12 was simple, just addition repeated through a list. Overall, a decent end to the project. And as always, my methods are further explained in my comments.